LayoutTests/accessibility/ios-simulator/textentry-traits.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script>
var successfullyParsed = false;
</script>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body id="body">

<input type="text" id="text1">
<input type="search" id="text2">
<textarea id="text3"></textarea>

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that input type: time is accessible on iOS.");

    if (window.accessibilityController) {

        // Regular text input should not be a search field or text area.
        shouldBeFalse("accessibilityController.accessibleElementById('text1').isSearchField || accessibilityController.accessibleElementById('text1').isTextArea");

        shouldBeTrue("accessibilityController.accessibleElementById('text2').isSearchField");
        shouldBeTrue("accessibilityController.accessibleElementById('text3').isTextArea");
    }

    successfullyParsed = true;
</script>

<script src="../../resources/js-test-post.js"></script>
</body>
</html>
