LayoutTests/accessibility/ios-simulator/textentry-traits-expected.txt


This tests that input type: time is accessible on iOS.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS accessibilityController.accessibleElementById('text1').isSearchField || accessibilityController.accessibleElementById('text1').isTextArea is false
PASS accessibilityController.accessibleElementById('text2').isSearchField is true
PASS accessibilityController.accessibleElementById('text3').isTextArea is true
PASS successfullyParsed is true

TEST COMPLETE
